latest from coda-oss

+++ externals/coda-oss/modules/c++/sys/source/OSWin32.cpp
@@ -205,7 +205,7 @@ sys::Off_T sys::OSWin32::getLastModifiedTime(const std::string& path) const
 
 void sys::OSWin32::millisleep(int milliseconds) const
 {
-    Sleep(milliseconds);
+    Sleep(static_cast<DWORD>(milliseconds));
 }
 
 std::string sys::OSWin32::getDSOSuffix() const
